Fixes issue 931, weight of variant should be float
Example in Shopify/shopify had an integer (100) for weigth thus wrapper
generation inferred Integer.  Changed the example to 100.0 to correct it
in the generated wrappers.

+++ lib/shopify_api/rest/resources/2021_07/variant.rb
@@ -36,7 +36,7 @@ module ShopifyAPI
       @taxable = T.let(nil, T.nilable(T::Boolean))
       @title = T.let(nil, T.nilable(String))
       @updated_at = T.let(nil, T.nilable(String))
-      @weight = T.let(nil, T.nilable(Integer))
+      @weight = T.let(nil, T.nilable(Float))
       @weight_unit = T.let(nil, T.nilable(String))
     end
 
@@ -103,7 +103,7 @@ module ShopifyAPI
     attr_reader :title
     sig { returns(T.nilable(String)) }
     attr_reader :updated_at
-    sig { returns(T.nilable(Integer)) }
+    sig { returns(T.nilable(Float)) }
     attr_reader :weight
     sig { returns(T.nilable(String)) }
     attr_reader :weight_unit
